Fixes #32000 - Remove legacy export/import API endpoints
