+++ locale/katello.pot
@@ -1225,12 +1225,6 @@ msgid ""
 "p 2 repositories"
 msgstr ""
 
-#: ../app/controllers/katello/api/v2/content_view_versions_controller.rb:95
-msgid ""
-"Invalid usage for Pulp 3 repositories. Use hammer content-export for Yum repos"
-"itories"
-msgstr ""
-
 #: ../app/controllers/katello/api/v2/content_view_versions_controller.rb:100
 #: ../app/controllers/katello/api/v2/repositories_controller.rb:314
 msgid "ISO export must be enabled when specifying ISO size"
