+++ test/actions/katello/content_view_version/import_test.rb
@@ -47,13 +47,11 @@ module ::Actions::Katello::ContentViewVersion
     def setup_proxy
       proxy = FactoryBot.create(:smart_proxy, :default_smart_proxy, :with_pulp3)
       SmartProxy.any_instance.stubs(:pulp_primary).returns(proxy)
-      SmartProxy.any_instance.stubs(:pulp_primary!).returns(proxy)
       proxy.smart_proxy_features.where(:feature_id => Feature.find_by(:name => SmartProxy::PULP_FEATURE)).delete_all
     end
 
     before do
       set_user
-      SmartProxy.any_instance.stubs(:ping_pulp).returns({})
       SmartProxy.any_instance.stubs(:ping_pulp3).returns({})
       SmartProxy.any_instance.stubs(:pulp3_configuration).returns(nil)
       ::Katello::Pulp3::Api::ContentGuard.any_instance.stubs(:list).returns(nil)
