+++ test/controllers/api/v2/content_views_controller_test.rb
@@ -91,22 +91,13 @@ module Katello
       assert_template 'katello/api/v2/common/create'
     end
 
-    def test_create_import_only_pulp3
+    def test_create_import
       FactoryBot.create(:smart_proxy, :default_smart_proxy, :with_pulp3)
 
       post :create, params: { :name => "My View", :label => "My_View", :description => "Cool",
                               :organization_id => @organization.id, :import_only => true }
 
       assert_response :success
-    end
-
-    def test_create_import_only_pulp2
-      FactoryBot.create(:smart_proxy, :default_smart_proxy)
-
-      post :create, params: { :name => "My View", :label => "My_View", :description => "Cool",
-                              :organization_id => @organization.id, :import_only => true }
-
-      assert_response :bad_request
     end
 
     def test_publish_with_version_params
